Numbering the ToDo list for referencing later
Potential I may add to the ToDo list. To this end, I'm labeling them with decimals so I can add in-between elements: like 2.1. I could just append to the list, but I'm planning on working on the list from top to bottom, and would like to keep new items near similar work.