Adding randomness and some game functionality
(1.0) For the most part the random generation of the final list is implemented in final_list.rs.

Added some logic for (3.0) checking the user's guess, and the proper ordering of the Pico and Fermis for displaying the user's result later (4.0). I will try to separate the display from the game data as coding progresses to make room for potential GUI building in the future.