Added basic user input
Game can now be played to completion with the addition of user input (2.0). Some error checking is made but panic! macros should be avoided in the future. Need to do some output formatting, turns out it will probably be impossible to get user input and then output something on the same line.
Additionally, testing usage of my GPG key and the Intellij IDE.